fix course views and controller

+++ ContosoUniversityTARpe21/Controllers/CoursesController.cs
@@ -2,14 +2,13 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using ContosoUniversityTARpe21.Data;
+using ContosoUniversityTARpe21.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 
-using ContosoUniversityTARpe21.Data;
-using ContosoUniversityTARpe21.Models;
-
-namespace ContosoUniversity.Controllers
+namespace UniversityRemusLuht.Controllers
 {
     public class CoursesController : Controller
     {
@@ -20,26 +19,23 @@ namespace ContosoUniversity.Controllers
             _context = context;
         }
 
-        //get index
+        // GET: Courses
         public async Task<IActionResult> Index()
         {
-            var courses = _context.Courses
-                .Include(c => c.Department)
-                .AsNoTracking();
-            return View(await courses.ToListAsync());
+            var schoolContext = _context.Courses.Include(c => c.Department);
+            return View(await schoolContext.ToListAsync());
         }
 
-        //get details
+        // GET: Courses/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Courses == null)
             {
                 return NotFound();
             }
 
             var course = await _context.Courses
                 .Include(c => c.Department)
-                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.CourseID == id);
             if (course == null)
             {
@@ -49,100 +45,95 @@ namespace ContosoUniversity.Controllers
             return View(course);
         }
 
-        //get create
+        // GET: Courses/Create
         public IActionResult Create()
         {
-            PopulateDepartmentsDropDownList();
+            ViewData["DepartmentID"] = new SelectList(_context.Departments, "DepartmentID", "Name");
             return View();
         }
 
-        //post create
+        // POST: Courses/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("CourseID,Credits,DepartmentID,Title")] Course course)
+        public async Task<IActionResult> Create([Bind("CourseID,Title,Credits,DepartmentID")] Course course)
         {
+            ModelState.Remove("Enrollments");
+            ModelState.Remove("CourseAssignments");
+            ModelState.Remove("Department");
             if (ModelState.IsValid)
             {
                 _context.Add(course);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            PopulateDepartmentsDropDownList(course.DepartmentID);
+            ViewData["DepartmentID"] = new SelectList(_context.Departments, "DepartmentID", "Name", course.DepartmentID);
             return View(course);
         }
 
-        //get edit
+        // GET: Courses/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Courses == null)
             {
                 return NotFound();
             }
 
-            var course = await _context.Courses
-                .AsNoTracking()
-                .FirstOrDefaultAsync(m => m.CourseID == id);
+            var course = await _context.Courses.FindAsync(id);
             if (course == null)
             {
                 return NotFound();
             }
-            PopulateDepartmentsDropDownList(course.DepartmentID);
+            ViewData["DepartmentID"] = new SelectList(_context.Departments, "DepartmentID", "Name", course.DepartmentID);
             return View(course);
         }
 
-        //post edit
+        // POST: Courses/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> EditPost(int? id)
+        public async Task<IActionResult> Edit(int id, [Bind("CourseID,Title,Credits,DepartmentID")] Course course)
         {
-            if (id == null)
+            ModelState.Remove("Enrollments");
+            ModelState.Remove("CourseAssignments");
+            ModelState.Remove("Department");
+            if (id != course.CourseID)
             {
                 return NotFound();
             }
 
-            var courseToUpdate = await _context.Courses
-                .FirstOrDefaultAsync(c => c.CourseID == id);
-
-            if (await TryUpdateModelAsync<Course>(courseToUpdate,
-                "",
-                c => c.Credits, c => c.DepartmentID, c => c.Title))
+            if (ModelState.IsValid)
             {
                 try
                 {
+                    _context.Update(course);
                     await _context.SaveChangesAsync();
-                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateException /* ex */)
+                catch (DbUpdateConcurrencyException)
                 {
-                    //Log the error (uncomment ex variable name and write a log.)
-                    ModelState.AddModelError("", "Unable to save changes. " +
-                        "Try again, and if the problem persists, " +
-                        "see your system administrator.");
+                    if (!CourseExists(course.CourseID))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
                 }
+                return RedirectToAction(nameof(Index));
             }
-            PopulateDepartmentsDropDownList(courseToUpdate.DepartmentID);
-            return View(courseToUpdate);
+            ViewData["DepartmentID"] = new SelectList(_context.Departments, "DepartmentID", "Name", course.DepartmentID);
+            return View(course);
         }
 
-        private void PopulateDepartmentsDropDownList(object selectedDepartment = null)
-        {
-            var departmentsQuery = from d in _context.Departments
-                                   orderby d.Name
-                                   select d;
-            ViewBag.DepartmentID = new SelectList(departmentsQuery.AsNoTracking(), "DepartmentID", "Name", selectedDepartment);
-        }
-
-        //get Delete
+        // GET: Courses/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Courses == null)
             {
                 return NotFound();
             }
 
             var course = await _context.Courses
                 .Include(c => c.Department)
-                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.CourseID == id);
             if (course == null)
             {
@@ -152,38 +143,28 @@ namespace ContosoUniversity.Controllers
             return View(course);
         }
 
-        //
-        [HttpPost, ActionName("Delete")]
+        // POST: Courses/Delete/5
+        [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> Delete(int id)
         {
+            if (_context.Courses == null)
+            {
+                return Problem("Entity set 'SchoolContext.Courses'  is null.");
+            }
             var course = await _context.Courses.FindAsync(id);
-            _context.Courses.Remove(course);
+            if (course != null)
+            {
+                _context.Courses.Remove(course);
+            }
+
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
-        public IActionResult UpdateCourseCredits()
-        {
-            return View();
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> UpdateCourseCredits(int? multiplier)
-        {
-            if (multiplier != null)
-            {
-                ViewData["RowsAffected"] =
-                    await _context.Database.ExecuteSqlRawAsync(
-                        "UPDATE Course SET Credits = Credits * {0}",
-                        parameters: multiplier);
-            }
-            return View();
-        }
-
         private bool CourseExists(int id)
         {
-            return _context.Courses.Any(e => e.CourseID == id);
+            return (_context.Courses?.Any(e => e.CourseID == id)).GetValueOrDefault();
         }
     }
 }
